Added comments and first steps for RicartAgrawalas
I've added some comments to help clarify what part of the protocols is
happening and where, and set up some of the stuff for implementing
Ricart and Agrawala's protocol.

Because some of the actions that need to be taken for certain parts
varies depending on the protocol (namely Release() and the processing of
incoming messages), should we have an enum that sets up which protocol
we're using, and pass that into the ServerThreads on startup?

We'll also need to add a list for deferred incoming requests that the
ServerThreads can push to.

Lastly, can someone update the readme with instructions for compiling?
I'm not sure which class to run or anything.